% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{An example data frame}
\format{
A data frame with parameters needed to diagnose MAFLD:
\describe{
  \item{CAP}{Controlled Attenuation Parameter in dB/m}
  \item{BMI}{Body mass index in kg/m2}
  \item{Race}{either Caucasians or Asians}
  \item{Gender}{Gender in Male or Female}
  \item{WC}{Waist Circumference in cm}
  \item{FPG}{Fasting plasma glucose in mg/dL}
  \item{HbA1C}{Glycated Hemoglobin (in\%)}
  \item{SBP}{Systolic BP in mm Hg}
  \item{DBP}{Diastolic BP in mm Hg}
  \item{HDL}{High Density Lipoprotein Cholesterol in mg/dL}
  \item{TG}{Triglycerides in mg/dL}
  \item{HOMAIR}{Homeostatis Model Assessment of Inuslin Resistance}
  \item{HSCRP}{High Sensitive C reative Protein}
}
}
\usage{
x
}
\description{
A data frame with exact column names as specified
}
\keyword{datasets}
